overlap try1

+++ rdma/RDMAStack.h
@@ -41,7 +41,7 @@ class RDMADispatcher {
   std::thread t;
   std::thread memory_t;
   int notify_malloc = -1;
-  std::atomic<uint64_t> wait_to_add = {0};
+  std::atomic<int> wait_to_add = {0};
 
   CephContext *cct;
   Infiniband::CompletionQueue* tx_cq = nullptr;
